refact(common): binomial coefficients
Binomial coefficients were stored in type int.
Since at large simulations overflow can happen, int64_t typed binomial coefficient was added.

The templated version can be used for larger numbers later as well.

+++ piquassoboost/sampling/source/GlynnPermanentCalculatorRepeated.h
@@ -127,7 +127,7 @@ void IterateOverDeltas(
     matrix_type& colSum,
     int sign,
     size_t index_min,
-    int currentMultiplicity
+    int64_t currentMultiplicity
 );
